Update 27.04.18 work

+++ src/GURPS/Space/Ship/AntimatterEngine.java
@@ -2,7 +2,7 @@ package GURPS.Space.Ship;
 
 public class AntimatterEngine extends Engine {
 
-    public AntimatterEngine (int size, long cost, byte type, String name, double acceleration, double deltaV, byte fuelType, int workspace, int techLvl) {
+    public AntimatterEngine (int size, long cost, byte type, String name, double acceleration, double deltaV, short fuelType, int workspace, int techLvl) {
         this.size=size;
         this.cost=cost;
         this.type=type;
