+++ src/GURPS/Space/Ship/ChemRockEngine.java
@@ -2,7 +2,7 @@ package GURPS.Space.Ship;
 
 public class ChemRockEngine extends Engine {
 
-    public ChemRockEngine (int size, long cost, byte type, String name, boolean violate, double acceleration, double deltaV, byte fuelType, int workspace, int techLvl) {
+    public ChemRockEngine (int size, long cost, byte type, String name, boolean violate, double acceleration, double deltaV, short fuelType, int workspace, int techLvl) {
         this.size=size;
         this.cost=cost;
         this.type=type;
