+++ src/GURPS/Space/Ship/ElecRockEngine.java
@@ -4,7 +4,7 @@ public class ElecRockEngine extends Engine {
 
     boolean highEnergy = true;
 
-    public ElecRockEngine (int size, long cost, byte type, String name, double acceleration, double deltaV, byte fuelType, int workspace, int techLvl) {
+    public ElecRockEngine (int size, long cost, byte type, String name, double acceleration, double deltaV, short fuelType, int workspace, int techLvl) {
         this.size=size;
         this.cost=cost;
         this.type=type;
