+++ src/GURPS/Space/Ship/FusionRockEngine.java
@@ -2,7 +2,7 @@ package GURPS.Space.Ship;
 
 public class FusionRockEngine extends Engine {
 
-    public FusionRockEngine (int size, long cost, byte type, String name, double acceleration, double deltaV, byte fuelType, int workspace, int techLvl) {
+    public FusionRockEngine (int size, long cost, byte type, String name, double acceleration, double deltaV, short fuelType, int workspace, int techLvl) {
         this.size=size;
         this.cost=cost;
         this.type=type;
