+++ src/GURPS/Space/Ship/NuclPulseRockEngine.java
@@ -2,7 +2,7 @@ package GURPS.Space.Ship;
 
 public class NuclPulseRockEngine extends Engine {
 
-    public NuclPulseRockEngine (int size, long cost, byte type, String name, double acceleration, double deltaV, byte fuelType, int workspace, int techLvl) {
+    public NuclPulseRockEngine (int size, long cost, byte type, String name, double acceleration, double deltaV, short fuelType, int workspace, int techLvl) {
         this.size=size;
         this.cost=cost;
         this.type=type;
